inference/predict.py
import os
import torch
import numpy as np
from pathlib import Path

class OCPPredictor:
    def __init__(self, model_path: str = None):
        """
        初始化预测器
        Args:
            model_path: 预训练模型路径，如果为None则使用默认路径
        """
        if model_path is None:
            model_path = os.path.join(
                Path(__file__).parent.parent,
                'checkpoints',
                'ocp_model_latest.pth'
            )
        
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = self._load_model(model_path)
        
    def _load_model(self, model_path: str):
        """
        加载预训练模型
        Args:
            model_path: 模型路径
        Returns:
            加载好的模型
        """
        if not os.path.exists(model_path):
            raise FileNotFoundError(f"Model file not found at {model_path}")
            
        # TODO: 从OCPNet导入具体的模型类
        # from OCPNet.model import OCPModel
        # model = OCPModel()
        # model.load_state_dict(torch.load(model_path, map_location=self.device))
        # model.to(self.device)
        # model.eval()
        # return model
        pass
        
    def predict(self, input_data: np.ndarray):
        """
        进行预测
        Args:
            input_data: 输入数据，numpy数组
        Returns:
            预测结果
        """
        with torch.no_grad():
            # 数据预处理
            input_tensor = torch.from_numpy(input_data).float().to(self.device)
            
            # 模型预测
            output = self.model(input_tensor)
            
            # 后处理
            result = output.cpu().numpy()
            
            return result
            
    def batch_predict(self, input_batch: np.ndarray):
        """
        批量预测
        Args:
            input_batch: 批量输入数据
        Returns:
            批量预测结果
        """
        results = []
        for data in input_batch:
            result = self.predict(data)
            results.append(result)
        return np.array(results)

if __name__ == "__main__":
    # 使用示例
    predictor = OCPPredictor()
    
    # 假设的输入数据
    sample_input = np.random.randn(10, 10)  # 替换为实际的输入数据格式
    
    # 单个预测
    result = predictor.predict(sample_input)
    print("预测结果:", result.shape)
    
    # 批量预测
    batch_input = np.random.randn(5, 10, 10)  # 5个样本
    batch_results = predictor.batch_predict(batch_input)
    print("批量预测结果:", batch_results.shape) 