requirements.txt
# Basic scientific computing and data processing
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
xarray>=0.20.0

# Geospatial data processing
rasterio>=1.2.0
netCDF4>=1.5.0

# Data visualization
matplotlib>=3.4.0
pillow>=8.0.0

# Marine data acquisition
copernicusmarine>=1.0.0

# Other utilities
requests>=2.26.0